desktop_messaging_app/requirements.txt
cryptography==42.0.8
Flask==2.3.2
Flask_SocketIO==5.3.7
pycryptodome==3.21.0
pymongo==4.10.1
pywebview==5.2
Werkzeug==2.3.4
